+++ operator-common/src/main/java/io/strimzi/operator/common/AdminClientProvider.java
@@ -7,6 +7,8 @@ package io.strimzi.operator.common;
 import io.fabric8.kubernetes.api.model.Secret;
 import org.apache.kafka.clients.admin.Admin;
 
+import java.util.Properties;
+
 /**
  * Interface to be implemented for returning an instance of Kafka Admin interface
  */
@@ -22,4 +24,17 @@ public interface AdminClientProvider {
      * @return Instance of Kafka Admin interface
      */
     Admin createAdminClient(String bootstrapHostnames, Secret clusterCaCertSecret, Secret keyCertSecret, String keyCertName);
+
+    /**
+     * Create a Kafka Admin interface instance
+     *
+     * @param bootstrapHostnames Kafka hostname to connect to for administration operations
+     * @param clusterCaCertSecret Secret containing the cluster CA certificate for TLS encryption
+     * @param keyCertSecret Secret containing keystore for TLS client authentication
+     * @param keyCertName Key inside the keyCertSecret for getting the keystore and the corresponding password
+     * @param config Additional configuration for the Kafka Admin Client
+     *
+     * @return Instance of Kafka Admin interface
+     */
+    Admin createAdminClient(String bootstrapHostnames, Secret clusterCaCertSecret, Secret keyCertSecret, String keyCertName, Properties config);
 }