+++ user-operator/src/main/java/io/strimzi/operator/user/UserOperatorConfig.java
@@ -8,9 +8,12 @@ import io.strimzi.api.kafka.model.CertificateAuthority;
 import io.strimzi.operator.common.InvalidConfigurationException;
 import io.strimzi.operator.common.model.Labels;
 
+import java.io.IOException;
+import java.io.StringReader;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.Properties;
 
 /**
  * Cluster Operator configuration
@@ -33,6 +36,7 @@ public class UserOperatorConfig {
     public static final String STRIMZI_KRAFT_ENABLED = "STRIMZI_KRAFT_ENABLED";
     public static final String STRIMZI_SCRAM_SHA_PASSWORD_LENGTH = "STRIMZI_SCRAM_SHA_PASSWORD_LENGTH";
     public static final String STRIMZI_MAINTENANCE_TIME_WINDOWS = "STRIMZI_MAINTENANCE_TIME_WINDOWS";
+    public static final String STRIMZI_KAFKA_ADMIN_CLIENT_CONFIGURATION = "STRIMZI_KAFKA_ADMIN_CLIENT_CONFIGURATION";
 
     public static final long DEFAULT_FULL_RECONCILIATION_INTERVAL_MS = 120_000;
     public static final String DEFAULT_KAFKA_BOOTSTRAP_SERVERS = "localhost:9091";
@@ -59,6 +63,7 @@ public class UserOperatorConfig {
     private final boolean kraftEnabled;
     private final int scramPasswordLength;
     private final List<String> maintenanceWindows;
+    private final Properties kafkaAdminClientConfiguration;
 
     /**
      * Constructor
@@ -79,6 +84,7 @@ public class UserOperatorConfig {
      * @param clientsCaRenewalDays How long before the certificate expiration should the user certificate be renewed
      * @param scramPasswordLength Length used for the Scram-Sha Password
      * @param maintenanceWindows Lit of maintenance windows
+     * @param kafkaAdminClientConfiguration Additional configuration for the Kafka Admin Client
      */
     @SuppressWarnings({"checkstyle:ParameterNumber"})
     public UserOperatorConfig(String namespace,
@@ -96,7 +102,9 @@ public class UserOperatorConfig {
                               int clientsCaValidityDays,
                               int clientsCaRenewalDays,
                               int scramPasswordLength,
-                              List<String> maintenanceWindows) {
+                              List<String> maintenanceWindows,
+                              Properties kafkaAdminClientConfiguration
+    ) {
         this.namespace = namespace;
         this.reconciliationIntervalMs = reconciliationIntervalMs;
         this.kafkaBootstrapServers = kafkaBootstrapServers;
@@ -113,6 +121,7 @@ public class UserOperatorConfig {
         this.clientsCaRenewalDays = clientsCaRenewalDays;
         this.scramPasswordLength = scramPasswordLength;
         this.maintenanceWindows = maintenanceWindows;
+        this.kafkaAdminClientConfiguration = kafkaAdminClientConfiguration;
     }
 
     /**
@@ -186,10 +195,33 @@ public class UserOperatorConfig {
 
         List<String> maintenanceWindows = parseMaintenanceTimeWindows(map.get(UserOperatorConfig.STRIMZI_MAINTENANCE_TIME_WINDOWS));
 
+        Properties kafkaAdminClientConfiguration = parseKafkaAdminClientConfiguration(map.get(UserOperatorConfig.STRIMZI_KAFKA_ADMIN_CLIENT_CONFIGURATION));
+
         return new UserOperatorConfig(namespace, reconciliationInterval, kafkaBootstrapServers, labels,
                 caCertSecretName, caKeySecretName, clusterCaCertSecretName, euoKeySecretName, caNamespace, secretPrefix,
                 aclsAdminApiSupported, kraftEnabled, clientsCaValidityDays, clientsCaRenewalDays,
-                scramPasswordLength, maintenanceWindows);
+                scramPasswordLength, maintenanceWindows, kafkaAdminClientConfiguration);
+    }
+
+    /**
+     * Parse the Kafka Admin Client configuration from the environment variable
+     *
+     * @param configuration The configuration from the environment variable. Null if no configuration is set.
+     *
+     * @return  The properties object with the configuration
+     */
+    /* test */ static Properties parseKafkaAdminClientConfiguration(String configuration) {
+        Properties kafkaAdminClientConfiguration = new Properties();
+
+        if (configuration != null)   {
+            try {
+                kafkaAdminClientConfiguration.load(new StringReader(configuration));
+            } catch (IOException | IllegalArgumentException e)   {
+                throw new InvalidConfigurationException("Failed to parse " + UserOperatorConfig.STRIMZI_KAFKA_ADMIN_CLIENT_CONFIGURATION + " configuration", e);
+            }
+        }
+
+        return kafkaAdminClientConfiguration;
     }
 
     /**
@@ -359,6 +391,13 @@ public class UserOperatorConfig {
         return maintenanceWindows;
     }
 
+    /**
+     * @return Properties object with the user-supplied configuration for the Kafka Admin Client
+     */
+    public Properties getKafkaAdminClientConfiguration() {
+        return kafkaAdminClientConfiguration;
+    }
+
     @Override
     public String toString() {
         return "ClusterOperatorConfig(" +
@@ -377,6 +416,7 @@ public class UserOperatorConfig {
                 ",clientsCaRenewalDays=" + clientsCaRenewalDays +
                 ",scramPasswordLength=" + scramPasswordLength +
                 ",maintenanceWindows=" + maintenanceWindows +
+                ",kafkaAdminClientConfiguration=" + kafkaAdminClientConfiguration +
                 ")";
     }
 }