+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Use CustomResource existing spec and status to fix Quarkus native build's serialization
 * Update JMX Exporter to version 0.17.0
 * Operator emits Kafka events to explain why it restarted a Kafka broker
+* Better configurability of the Kafka Admin client in the User Operator
 
 ## 0.29.0